playground/studio-v2/backend/requirements.txt
fastapi
uvicorn
python-multipart
pdf2image
google-cloud-aiplatform>=1.38.0
sqlalchemy
alembic
ultralytics
loguru
python-dotenv==1.2.1
